+++ mbo/types/no_destruct_test.cc
@@ -64,8 +64,8 @@ TEST_F(NoDestructTest, ClangCheck) {
 #endif  // _LIBCPP_STD_VER
 
 TEST_F(NoDestructTest, Test) {
-    const auto expected_simple = Conditional(kStructNameSupport, "{a: 25, b: 42}", "{25, 42}");
-    const auto expected_string = Conditional(kStructNameSupport, R"({a: "25", b: "42"})", R"({"25", "42"})");
+    const auto expected_simple = Conditional(kStructNameSupport, "{.a: 25, .b: 42}", "{25, 42}");
+    const auto expected_string = Conditional(kStructNameSupport, R"({.a: "25", .b: "42"})", R"({"25", "42"})");
     EXPECT_THAT(kTestSimple.Get().ToString(), expected_simple);
     EXPECT_THAT((*kTestSimple).ToString(), expected_simple);
     EXPECT_THAT(kTestSimple->ToString(), expected_simple);